add IReportRenderer
